% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-equal.R
\name{design_equal}
\alias{design_equal}
\title{Test if two designs are equivalent}
\usage{
design_equal(x, y, ...)
}
\arguments{
\item{x, y}{A data frame with the design of experiment table}

\item{...}{A series of name-value pair specifying which columns in \code{x} and
\code{y} are equivalent. The name should be the name of the column in \code{x} and
value should be the name of column in \code{y}.}
}
\description{
Given two data frames that contain the design of experiment output, test
to see if the design is equivalent.
}
\examples{
\dontrun{
library(edibble)
des1 <- design() \%>\%
  set_units(plot = 30) \%>\%
  set_trts(trt = 5) \%>\%
  allot_table(trt ~ plot)

des2 <- design() \%>\%
  set_units(subject = 30) \%>\%
  set_trts(vaccine = 5) \%>\%
  allot_table(vaccine ~ subject)

design_equal(des1, des2,
             plot = subject, trt = vaccine)
}
}
