% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasted-package.R
\docType{package}
\name{tasted-package}
\alias{tasted}
\alias{tasted-package}
\title{tasted: A testing framework for the table of designed experiments}
\description{
Testing for functions that produce an experimental design are difficult because of the stochastic nature of the output. This package provides a set of functions that provide checks of properties of the experimental design output.
}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID})

}
\keyword{internal}
